ignore aready registered sort certs
