+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/dccreissuance/ui/consent/acccerts/DccReissuanceAccCertsViewModel.kt
@@ -1,6 +1,5 @@
 package de.rki.coronawarnapp.dccreissuance.ui.consent.acccerts
 
-import androidx.annotation.VisibleForTesting
 import androidx.lifecycle.LiveData
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
@@ -8,14 +7,9 @@ import dagger.assisted.AssistedInject
 import de.rki.coronawarnapp.ccl.dccwalletinfo.model.CertificateReissuance
 import de.rki.coronawarnapp.covidcertificate.common.certificate.DccQrCodeExtractor
 import de.rki.coronawarnapp.covidcertificate.common.certificate.DccV1Parser
-import de.rki.coronawarnapp.covidcertificate.common.qrcode.DccQrCode
 import de.rki.coronawarnapp.covidcertificate.person.core.PersonCertificatesProvider
-import de.rki.coronawarnapp.covidcertificate.recovery.core.qrcode.RecoveryCertificateQRCode
-import de.rki.coronawarnapp.covidcertificate.test.core.qrcode.TestCertificateQRCode
-import de.rki.coronawarnapp.covidcertificate.vaccination.core.qrcode.VaccinationCertificateQRCode
 import de.rki.coronawarnapp.dccreissuance.ui.consent.DccReissuanceCertificateCard
 import de.rki.coronawarnapp.dccreissuance.ui.consent.DccReissuanceItem
-import de.rki.coronawarnapp.util.TimeAndDateExtensions.toLocalDateUserTz
 import de.rki.coronawarnapp.util.coroutine.DispatcherProvider
 import de.rki.coronawarnapp.util.viewmodel.CWAViewModel
 import de.rki.coronawarnapp.util.viewmodel.CWAViewModelFactory
@@ -56,15 +50,5 @@ class DccReissuanceAccCertsViewModel @AssistedInject constructor(
         fun create(
             personIdentifierCode: String
         ): DccReissuanceAccCertsViewModel
-    }
-}
-
-@VisibleForTesting
-internal fun List<DccQrCode>.sort() = sortedByDescending {
-    when (it) {
-        is TestCertificateQRCode -> it.data.certificate.test.sampleCollectedAt?.toLocalDateUserTz()
-        is VaccinationCertificateQRCode -> it.data.certificate.vaccination.vaccinatedOn
-        is RecoveryCertificateQRCode -> it.data.certificate.recovery.testedPositiveOn
-        else -> null
     }
 }
